+++ src/main/java/com/miniim/domain/mapper/FriendRelationMapper.java
@@ -1,0 +1,7 @@
+package com.miniim.domain.mapper;
+
+import com.baomidou.mybatisplus.core.mapper.BaseMapper;
+import com.miniim.domain.entity.FriendRelationEntity;
+
+public interface FriendRelationMapper extends BaseMapper<FriendRelationEntity> {
+}