+++ src/main/java/com/miniim/domain/service/MessageService.java
@@ -4,5 +4,5 @@ import com.baomidou.mybatisplus.extension.service.IService;
 import com.miniim.domain.entity.MessageEntity;
 
 public interface MessageService extends IService<MessageEntity> {
-    Long getOrCreateSingleChatId(Long fromUserId, Long toUserId);
+//    Long getOrCreateSingleChatId(Long fromUserId, Long toUserId);
 }